Abstract ActionButton implementation
