--- apps/website/src/components/CopyToClipboardButton.tsx
@@ -1,45 +1,0 @@
-import { useCallback } from "react";
-
-import { classes } from "@/utils/classes";
-
-import {
-  type CopyToClipboardOptions,
-  useCopyToClipboard,
-} from "@/hooks/clipboard";
-
-import IconClipboard from "@/icons/IconClipboard";
-
-type CopyToClipboardButtonProps = {
-  text: string;
-  options?: CopyToClipboardOptions;
-};
-
-const CopyToClipboardButton = ({
-  text,
-  options,
-}: CopyToClipboardButtonProps) => {
-  const { copy, status, statusText } = useCopyToClipboard(options);
-  const onClick = useCallback(async () => {
-    await copy(text);
-  }, [copy, text]);
-
-  return (
-    <div className="group relative inline-block">
-      <button onClick={onClick} title="Copy to clipboard">
-        <IconClipboard className="m-1 inline size-4 cursor-pointer text-alveus-green-400 group-hover:text-black" />
-      </button>
-
-      <span
-        className={classes(
-          "pointer-events-none absolute top-1/2 left-full z-10 -translate-y-1/2 rounded-md bg-alveus-green-900 px-1 text-white transition-opacity",
-          status === undefined ? "opacity-0" : "opacity-100",
-          "group-hover:opacity-100",
-        )}
-      >
-        {statusText}
-      </span>
-    </div>
-  );
-};
-
-export default CopyToClipboardButton;
